requirements.txt
# # Deep Learning Frameworks - GPU enabled versions
torch==2.1.0+cu121
torchvision==0.16.0+cu121
torchaudio==2.1.0+cu121

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1

# Data Processing
numpy>=1.23.5  # Updated for Python 3.11 compatibility
pandas==2.1.1
scikit-learn==1.3.1

# Visualization
matplotlib==3.8.0
seaborn==0.13.0

# Explainability
grad-cam==1.4.8

# Web Framework
streamlit==1.28.0

# Progress Bars
tqdm==4.66.1

# Utilities
PyYAML==6.0.1
kaggle==1.5.16

# Additional Dependencies
typing-extensions==4.15.0
MarkupSafe==2.1.5
filelock==3.19.1
fsspec==2023.9.0
jinja2==3.1.6
mpmath==1.3.0
networkx==3.5
sympy==1.13.1